UI: Move NoRowsOverlay into a separate folder
