+++ monkey/monkey_island/cc/next_ui/src/app/(protected)/plugins/_lib/no-rows-overlay/style.ts
